# docs/protocol/openapi.yaml
openapi: 3.1.0
info:
  title: QuanCard Server API
  version: 1.0.0
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
  description: 'HTTP API implemented by the server. All JSON request objects described here reject unknown properties.
    Routes without a requestBody do not inspect a parsed body. Binary revision uploads contain an opaque sync-v1
    envelope JSON object, not a JSON API wrapper. Base64 fields use the standard padded alphabet. Timestamps are
    Unix seconds.


    API requests require HTTPS and a Host matching the configured public origin, with an explicitly configured localhost
    development exception. GET /healthz bypasses these checks. Non-Bearer API writes require Origin matching the
    allowed origin and X-QuanCard-Client: web, including unauthenticated setup and login; /api/v1/pairings/claim
    is exempt. The Bearer-prefix exemption is a transport guard and does not grant authentication or access.


    Session cookies are host-only, Secure, HttpOnly and SameSite=Strict. Device bearer tokens begin with qcd_ and
    can only access manifest and revisions of their paired vault. A recognized qcd_ bearer credential takes precedence
    over a session cookie. API responses use Cache-Control: no-store. Errors use Error and HTTP 429 also includes
    Retry-After. HEAD counterparts are automatically exposed and return no body.


    When webRoot is configured, the server additionally serves deployment-dependent static files and returns its
...
servers:
...
paths:
...
  /api/v1/status:
    get:
      operationId: getStatus
      summary: Get server and setup status
      description: Public service metadata; no session required.
      security: []
      responses:
        '200':
          description: Current server status.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                - service
                - version
                - setupRequired
                - setupEnabled
                - protocol
                properties:
                  service:
                    type: string
                    const: quancard-server
                  version:
                    type: string
                    description: Configured server build version.
                  setupRequired:
                    type: boolean
                  setupEnabled:
                    type: boolean
                  protocol:
                    type: object
                    additionalProperties: false
                    required:
                    - auth
                    - sync
                    - envelope
                    properties:
                      auth:
                        type: integer
                        const: 1
                      sync:
                        type: integer
                        const: 1
                      envelope:
                        type: integer
                        const: 1
        '421':
          $ref: '#/components/responses/MisdirectedRequest'
        '426':
          $ref: '#/components/responses/HttpsRequired'
        '500':
          $ref: '#/components/responses/Internal'
    head:
      operationId: headGetStatus
      summary: 'Get headers: Get server and setup status'
      description: Automatically exposed by Fastify for this GET route. Uses the same handler and authentication,
        but suppresses response bodies, including Error bodies.
      security: []
      responses:
        '200':
          description: Current server status. HEAD responses have no body.
        '421':
          description: 'misdirectedRequest: Host does not match the configured public origin (except permitted local
            development requests). HEAD responses have no body.'
        '426':
          description: 'httpsRequired: HTTPS is required; only explicitly enabled localhost development is exempt.
            HEAD responses have no body.'
          headers:
            Connection:
              schema:
                type: string
                const: close
              description: The server closes the insecure connection.
        '500':
          description: 'internal: unexpected server failure. HEAD responses have no body.'
  /api/v1/setup:
    post:
      operationId: setupOwner
      summary: Create the initial owner
      description: Available only with a configured setup token and an empty user database. Limited to 10 attempts
        per minute per client.
      security: []
      parameters:
      - $ref: '#/components/parameters/Origin'
      - $ref: '#/components/parameters/WebClient'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetupRequest'
      responses:
        '201':
          description: Owner created; session issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
          headers:
            Set-Cookie:
              description: Sets __Host-qc_session with Path=/, HttpOnly, Secure and SameSite=Strict, no Domain,
                and the configured absolute-session Max-Age.
              schema:
                type: string
        '400':
          description: 'badRequest: malformed or unknown fields, invalid UUID/Base64 or account-key envelope. invalidUsername:
            invalid normalized username.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 'setupClosed: setup token is disabled or any user already exists; invalidSetupToken: incorrect
            setup token; crossOriginRejected: browser write headers failed.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 'usernameTaken: username already exists.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '421':
          $ref: '#/components/responses/MisdirectedRequest'
        '426':
          $ref: '#/components/responses/HttpsRequired'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
  /api/v1/invites:
    post:
      operationId: createInvite
      summary: Create a registration invite
      description: Owner session required. No request body is required or inspected; a parsed body is ignored.
      security:
      - session: []
      parameters:
      - $ref: '#/components/parameters/Origin'
      - $ref: '#/components/parameters/WebClient'
      responses:
        '201':
          description: Single-use invite created; expires after 7 days.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodeWithExpiry'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: 'forbidden: only the owner may create invites; crossOriginRejected: browser write headers
            failed.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '421':
          $ref: '#/components/responses/MisdirectedRequest'
        '426':
          $ref: '#/components/responses/HttpsRequired'
        '500':
          $ref: '#/components/responses/Internal'
  /api/v1/register:
    post:
      operationId: registerAccount
      summary: Register using an invite
      description: Creates a non-owner account with a single-use invite. Limited to 10 attempts per minute per client.
      security: []
      parameters:
      - $ref: '#/components/parameters/Origin'
      - $ref: '#/components/parameters/WebClient'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created, invite consumed and session issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
          headers:
            Set-Cookie:
              description: Sets __Host-qc_session with Path=/, HttpOnly, Secure and SameSite=Strict, no Domain,
                and the configured absolute-session Max-Age.
              schema:
                type: string
        '400':
          description: 'badRequest: malformed or unknown fields, invalid UUID/Base64 or account-key envelope. invalidUsername:
            invalid normalized username.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 'invalidInvite: unknown, used or expired invite; crossOriginRejected: browser write headers
            failed.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 'usernameTaken: username already exists.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '421':
          $ref: '#/components/responses/MisdirectedRequest'
        '426':
          $ref: '#/components/responses/HttpsRequired'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
  /api/v1/auth/prelogin:
    post:
      operationId: prelogin
      summary: Get account KDF parameters
      description: Unknown usernames receive a stable decoy salt without revealing account existence. Limited to
        10 attempts per minute per client.
      security: []
      parameters:
      - $ref: '#/components/parameters/Origin'
      - $ref: '#/components/parameters/WebClient'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreloginRequest'
      responses:
        '200':
          description: KDF parameters with the real or a stable decoy salt.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                - kdf
                properties:
                  kdf:
                    $ref: '#/components/schemas/Kdf'
        '400':
          description: 'badRequest: malformed or unknown fields; invalidUsername: invalid normalized username.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          $ref: '#/components/responses/CrossOriginRejected'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '421':
          $ref: '#/components/responses/MisdirectedRequest'
        '426':
          $ref: '#/components/responses/HttpsRequired'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
  /api/v1/auth/login:
    post:
      operationId: login
      summary: Authenticate and create a browser session
      description: The password remains client-side. Supply its derived authentication key and a second factor when
        TOTP is enabled. A string totp takes precedence over recoveryCode; factors are ignored when TOTP is disabled.
      security: []
      parameters:
      - $ref: '#/components/parameters/Origin'
      - $ref: '#/components/parameters/WebClient'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated account; session issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
          headers:
            Set-Cookie:
              description: Sets __Host-qc_session with Path=/, HttpOnly, Secure and SameSite=Strict, no Domain,
                and the configured absolute-session Max-Age.
              schema:
                type: string
        '400':
          description: 'badRequest: malformed or unknown fields or invalid authentication-key Base64; invalidUsername:
            invalid normalized username.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 'invalidCredentials: incorrect authentication key; secondFactorRequired: enabled TOTP requires
            a factor; invalidSecondFactor: invalid or replayed TOTP, or invalid/used recovery code.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          $ref: '#/components/responses/CrossOriginRejected'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '421':
          $ref: '#/components/responses/MisdirectedRequest'
        '426':
          $ref: '#/components/responses/HttpsRequired'
        '429':
          description: 'rateLimited: more than 10 login attempts per minute per client; accountLocked: account locked
            for 15 minutes after 10 failed authentication-key checks. The tenth failure still returns invalidCredentials;
            subsequent attempts during the lock return accountLocked.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          headers:
            Retry-After:
              description: Seconds until another attempt is allowed; also returned as retryAfterSeconds.
              schema:
                type: integer
                minimum: 1
        '500':
          $ref: '#/components/responses/Internal'
  /api/v1/auth/logout:
    post:
      operationId: logout
      summary: End the current browser session
      description: Succeeds without valid credentials. A device token is not revoked. No request body is required
        or inspected.
      security:
      - {}
      - session: []
      parameters:
      - $ref: '#/components/parameters/Origin'
      - $ref: '#/components/parameters/WebClient'
      responses:
        '204':
          description: Session revoked if present and cookie cleared; no response body.
          headers:
            Set-Cookie:
              description: Clears __Host-qc_session with Path=/, HttpOnly, Secure and SameSite=Strict.
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/CrossOriginRejected'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '421':
          $ref: '#/components/responses/MisdirectedRequest'
        '426':
          $ref: '#/components/responses/HttpsRequired'
        '500':
          $ref: '#/components/responses/Internal'
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: __Host-qc_session
      description: Browser session cookie. Device tokens cannot substitute for a session on account, vault-management
        or device-management endpoints.
    device:
      type: http
      scheme: bearer
      bearerFormat: qcd_<opaque token>
      description: Opaque device token issued by pairing claim. Only accepts manifest GET/HEAD, revisions GET/HEAD
        and revision PUT for its single paired vault.
...
  responses:
...
    MisdirectedRequest:
      description: 'misdirectedRequest: Host does not match the configured public origin (except permitted local
        development requests).'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...